🚚 move(ui): `testIds.ts`para pasta `utils`
